refactor login api route
