+++ main.go
@@ -26,8 +26,8 @@ func main() {
 	router.Use(middleware.Recovery("Something went wrong"))
 
 	router.POST("/api/v1/account/sign-up", api.SignUpHandler)
+	router.POST("/api/v1/account/login", api.LoginHandler)
 
-	router.POST("/api/v1/jwt/login", jwt.LoginWithJWTHandler)
 	router.POST("/api/v1/jwt/logout", jwt.Logout)
 	router.POST("/api/v1/jwt/forgot-password", jwt.ForgotPasswordHandler)
 	router.GET("/api/v1/jwt/auth-check", jwt.AuthCheckHandler)
